Wrap Inline tags
When an inline spanning element is of the form <element>word</element>
the default breaking algorithm in XOM's Serializer doesn't wrap before
the openning tag. 

So calculate the width of such elements, search for an available space
break, and break before if it's too wide.

The end result is that XOM seems to do the right thing anyway if there
are attributes present, though that seems a bit by accident.