Switch between preview and help
Create a placeholder for the help system, HelpWidget.

Add PreviewWidget and HelpWidget to a Notebook on the right hand side
of the PrimaryWindow. Hookup key handlers in PrimaryWindow to switch
between them. Tell EditorWidget to pass the characters through.

Move toggleFullscreen() to be a method on PrimaryWindow. It's not
really a UI wide thing as there's only one Window that can be full
screen.