Handle TextView scrolling
When the cursor is moved up above or down below the visible viewport,
react by passing the Allocation height + TreeIter rectangle height
through to ComponentEditorWidget's ensureVisible() method to change
the value property of the vertical scrollbar Adjustment, causing a
scoll to happen.

Add some lorem ipsum text to the example to force a larger canvas than
the available viewport.