Insert WORD JOINER before open brackets when laying out code text
Switch to using our new Pango Attributes API, which among other things
dramatically simplifies our rendering code here as we don't need a
second pass over the Extract to apply the Attributes.

This allows us to create Attirbutes as we go, meaning we can insert a
different number of characters to the width of the Span.

Change translate() to translateAndAppend(), returning the number of
characters added. Improve documentation here as this is a fairly
crucial method.