Initial implementation of splice UI
On pressing Insert within an EditorTextView, and selecting
"Preformatted", jump through many hoops to split the Segment in two,
inserting a new Segment, and then passing all this up to the
UserInterface so that it can propegate it back down through
PrimaryWindow and ComponentEditorWidget to create new Widgets as
necessary and repack them all.

Messy.