Add example document
Set Quack schema version to "0.1" and XML namespace to
"operationaldynamics.com/quack".

Add an example XML file conforming to the quack DTD as validated with
xmllint. This document is obviously the same content as the
tests/ExamplProgram.xml DocBook XML that we have been testing Quill
against, with the addition of endnotes and references so that we can
give the quoted passage proper full attribution.