Spell check on load
Initiate a spell check when each EditorTextView is loaded.

This required fixing a serious bug in UnicodeSpan whereby we were
returning the wrong codepoint when its getChar() implementation was
called. Fix this with tests in ValidateWordExtraction.

Write a new test ValidateSpellingOperations, which was where we
discovered the above bug was somehow related to word splitting.