Add text version of Iain M. Banks' essay on his Culture setting
Used under the following terms:

"Copyright 1994 Iain M Banks. Commercial use only by permission.
Other uses, distribution, reproduction, tearing to shreds etc are
freely encouraged provided the source is acknowledged."

The text is available in various locations around the internet, but
the citation for its original source is:

"This article was posted to newsgroup rec.arts.sf.written on 10 Aug
1994 on behalf of Iain M Banks by Ken MacLeod
<kenm@festival.ed.ac.uk>"

This text imported here is from what appears to be the author's blog,
http://blogs.myspace.com/index.cfm?fuseaction=blog.view&friendId=145352428&blogId=213851237