Use chapter title as window title
Set the window title based on the leading ComponentSegment's text,
which is the chapter title.

Necessitated making the EditorTextView's TextBuffer visible within the
package. That's not ideal. Hookup a handler to react to the content of
the editor changing and invoke the code to (re)set the window title.

Since we have the application name in the title, we don't use
"Untitled" for new documents.