Add an image of an ancient map
This image seems to be in various forms all over the net; a simple
search for "old map" turns up hundreds, and it was difficult to
ascertain the identity of the original copyright holder, if any. We
are using it here only for testing the relevant code paths, and will
not ship it as a part of the release.