Move Extract class out of the way.
Create new interface Extract, to be implemented by Node. This isn't
really that important, but it gives us a public face to our b-trees
that means we can swap out the implementation later if we wish, rather
than being tied to Node (which is anyway a bad name for something
public).

Pull SpanVisitor and CharacterVisitor out of Node.