Update Node's append() to semi-balance
While not a proper rebalancing, have the append() method create new
subtrees on the right if left is taller, with the result that we have
a far more complex tree during testing.

Correct numerous resultant failures in the getWordBoundaries*() logic.
Change these methods return -1 for "not found" so as to avoid
ambiguity with the boundary conditions of a ' ' existing at position
zero.