Use an HPaned to divide editor side and preview side
Swap HBox out for an HPaned. Set divider position to 680px. This seems
about right. We may need code in the future to persist the position.

Reset fonts to 11pt again, but use Condensed for fonts.serif making
the bold a bit nicer. This is really going to need to be dynamic.