Refactor UI to affect tree states
Remove use of Change objects and instead propegate updates up the
widget hierarchy, ending with the formation of a new Folio object.
Removes the undo()/redo() and apply()/reverse() methods.

This introduces a considerable loss in precision about where an edit
happened, meaning that we iterate over just about everything,
especially on undo/redo. This is expensive, but a work in progress,
and something we can improve once we reach correctness again.