Repair QuackLoader to generate immutable Segments
Implementation is a bit inefficient, in that we keep popping the
latest segment off the list and creating a new one. But we can improve
it later.

Move mutable image field in Segment to an immutable field, settable by
a new constructor in Segment used by ImageSegment's <init>() and
createSimilar() to stash image src attributes.