Repair undo/redo, again
Treat undo and redo as seperate but parallel code paths. This is
definitely sub-optimal, since we are duplicating code like mad.

Change ChangeStack to always have an element, so that undoing past the
beginning returns the first Folio, and redo()ing past the end returns
the last Folio, not null.