Make the background colour of the editing canvas white
Hide horizontal scrollbars when their PreformatEditorTextView is not
focused by drawing white on their space, and then set the backdrop of
the entire ComponentEditorWidget to white.

This solves the problem of the upper line of the slider appearing to
float as if it were an empty Block, at the cost of no visible
separator between blocks at all. Hm.