Move references to right side; swap accelerators
Move the references over to the right hand side of the PrimaryWindow.
While that's suboptimal for use when writing notes, it is still
ancillary lookaside information with respect to the main body of the
document, and we should be able to have the editor up when looking at
references. Not to mention some people will be working all in
references and not with endnotes at all, making it the same as notes
placement wise.

Reassign the accelerators. F4,F8 for references,endnotes wasn't
cutting it. Switch to F1... for editor side, and F5... for preview
side.

Live with Shift+F1 as help for now. It'll become Shift+F(n) for
context help in due course.