Stop Enchant crashing with mutliple upper case words
Enchant doesn't seem to handle it if there is a blank line in a
personal word list. So, although we were following the conventions
observed in ~/.config/enchant/en_CA.dic in our /tmp file, it crashed
Enchant. Great. Not doing that leading blank line seems to fix
problem.

Thanks to Serkan Kaba for helping with the diagnosis.